created Data Sections
Added new data sections to URLs so that data is sectioned off, rather one one giant page. This was in hopes to make the data a little less daunting.

New sections added: National, State, Data Tables, and About. The main page also loads the About page, or if a user types in a bad link.